User create a Sub Admin and give him access only to creates employees and creates multiple employees up-to-date. adminemployeronboarding adminsignupwithbasicdetail cucumberreport bulkemployeevalidation addcampdetailinemployee camp     Then    [Admin Page] Authorizer reviews the employee records they approve by the checker
ProcessSalaryEmployeeData
Employer downloads the salary file using the establishment ID, modifies the file and generates the SIF file, then submits it for approval. The admin checks the salary approval and verifies both the original file and the SIF file at the checker and authorizer levels.

allure reports

allure report complete